+++ db/Tables/PollingStates.sql
@@ -1,5 +1,8 @@
 CREATE TABLE [dbo].[PollingStates] (
   [Key] NVARCHAR(256) NOT NULL,
   [ETag] NVARCHAR(1024) NOT NULL,
+  [DateCreated] DATETIMEOFFSET NOT NULL,
+  [DateUpdated] DATETIMEOFFSET NOT NULL,
+  [DateStateChanged] DATETIMEOFFSET NOT NULL,
   CONSTRAINT [PK_PollingStates] PRIMARY KEY ([Key])
 );